+++ libs/angular-components/community/components/overlay/modal/modal.component.ts
@@ -1,5 +1,6 @@
 import {
   Component,
+  computed,
   inject,
   model,
   OnInit,
@@ -10,14 +11,15 @@ import { ModalHeaderComponent } from "./header/modal-header.component";
 import { ModalFooterComponent } from "./footer/modal-footer.component";
 import { ComponentInputs } from "@tehik-ee/tedi-angular/tedi";
 
-export type ModalSizes = "lg" | "md" | "sm";
-
 export type DialogData = ComponentInputs<
   ModalHeaderComponent & ModalFooterComponent
 > & {
-  maxWidth?: ModalSizes;
+  maxWidth?: ModalBreakpoint;
   variant?: "default" | "small";
 };
+
+export const modalBreakpoints = ["xs", "sm", "md", "lg", "xl"] as const;
+export type ModalBreakpoint = "xs" | "sm" | "md" | "lg" | "xl";
 
 @Component({
   selector: "tedi-modal",
@@ -26,20 +28,32 @@ export type DialogData = ComponentInputs<
   imports: [DialogModule],
   encapsulation: ViewEncapsulation.None,
   host: {
-    "[class.tedi-modal]": "true",
-    "[class.tedi-modal--lg]": "maxWidth() === 'lg'",
-    "[class.tedi-modal--md]": "maxWidth() === 'md'",
-    "[class.tedi-modal--sm]": "maxWidth() === 'sm'",
-    "[class.tedi-modal--small]": "variant() === 'small'",
+    "[class]": "hostClasses()",
   },
 })
 export class ModalComponent implements OnInit {
-  readonly maxWidth = model<ModalSizes>("sm");
+  readonly maxWidth = model<ModalBreakpoint>("sm");
   readonly variant = model<"default" | "small">("default");
 
   readonly dialogRef = inject(DialogRef, { optional: true });
   readonly dialogData = inject(DIALOG_DATA, {
     optional: true,
+  });
+
+  // host classes
+  readonly hostClasses = computed(() => {
+    const classes = ["tedi-modal"];
+
+    if (this.variant() === "small") {
+      classes.push("tedi-modal--padding-small");
+    }
+
+    for (const breakpoint of modalBreakpoints) {
+      if (this.maxWidth() === breakpoint) {
+        classes.push(`tedi-modal--width-${breakpoint}`);
+      }
+    }
+    return classes.join(" ");
   });
 
   ngOnInit(): void {