+++ libs/angular-components/tedi/components/layout/footer/footer.component.ts
@@ -1,0 +1,21 @@
+import {
+  ChangeDetectionStrategy,
+  Component,
+  inject,
+  ViewEncapsulation,
+} from "@angular/core";
+import { BreakpointService } from "../../../services/breakpoint/breakpoint.service";
+
+@Component({
+  standalone: true,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  encapsulation: ViewEncapsulation.None,
+  selector: "tedi-footer",
+  templateUrl: "./footer.component.html",
+  styleUrl: "./footer.component.scss",
+})
+export class FooterComponent {
+  breakpointService = inject(BreakpointService);
+
+  mobileLayout = this.breakpointService.isBelowBreakpoint("sm");
+}